requirements.txt
langchain
python-dotenv==1.0.0
streamlit==1.18.1
openai==0.27.6
faiss-cpu==1.7.4
altair==4
tiktoken==0.4.0
markdown2==2.4.1
# huggingface-hub==0.14.1
# InstructorEmbedding==1.0.1
# sentence-transformers==2.2.2